# api/openapi.yaml
openapi: 3.0.3
info:
  title: Alancoin API
  description: |
    Payment infrastructure for autonomous AI agents.
    
    Alancoin provides:
    - Agent registry and discovery
    - Session keys for bounded autonomy
    - Balance management and withdrawals
    - Reputation scoring
    - Webhooks for event notifications
    
    ## Authentication
    
    Most read endpoints are public. Write endpoints require an API key:
    
    ```
    Authorization: Bearer ap_sk_live_xxxxx
    ```
    
    API keys are issued on agent registration.
    
    ## Session Keys
    
    Session keys provide bounded autonomy - agents can transact within limits
    without human approval for each transaction.
    
    Flow:
    1. Generate ECDSA keypair client-side
    2. Register public key with limits (max per tx, per day, etc.)
    3. Sign transactions with private key
    4. Server validates signature and permissions, executes on-chain
    
  version: "0.1.0"
  contact:
    email: hello@alancoin.network
  license:
    name: MIT
servers:
  - url: https://api.alancoin.network
    description: Production
  - url: https://api.staging.alancoin.network
    description: Staging (Base Sepolia)
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /v1/agents/{address}/sessions:
    post:
      tags: [Session Keys]
      summary: Create session key
      security:
        - ApiKey: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [publicKey]
              properties:
                publicKey:
                  type: string
                expiresIn:
                  type: string
                  example: "24h"
                maxPerTransaction:
                  type: string
                  example: "1.00"
                maxPerDay:
                  type: string
                  example: "10.00"
      responses:
        '201':
          description: Session key created
  /v1/agents/{address}/sessions/{keyId}/transact:
    post:
      tags: [Session Keys]
      summary: Execute signed transaction
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
        - name: keyId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [to, amount, nonce, timestamp, signature]
              properties:
                to:
                  type: string
                amount:
                  type: string
                nonce:
                  type: integer
                timestamp:
                  type: integer
                signature:
                  type: string
      responses:
        '200':
          description: Transaction executed
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    enum: [approved, executed]
                  txHash:
                    type: string
  /v1/agents/{address}/webhooks:
    post:
      tags: [Webhooks]
      summary: Create webhook
      security:
        - ApiKey: []
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, events]
              properties:
                url:
                  type: string
                  format: uri
                events:
                  type: array
                  items:
                    type: string
                    enum:
                      - payment.received
                      - payment.sent
                      - session_key.used
                      - balance.deposit
                      - balance.withdraw
      responses:
        '201':
          description: Webhook created, secret returned
...
components:
  securitySchemes:
    ApiKey:
      type: http
      scheme: bearer
      description: API key from agent registration